smart_customer_service/requirements.txt
litellm>=1.31.0
fastapi>=0.111.0
uvicorn>=0.29.0
pydantic>=2.7.0
sqlalchemy>=2.0.0
python-multipart
python-dotenv
pandas>=2.0.0
scikit-learn>=1.4.0